backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
mysql-connector-python
python-multipart
PyJWT
passlib[bcrypt]
python-jose
python-dotenv
pydantic[email]